[Refactor] move smtp java sender package to infrascturcture
