+++ application/src/main/kotlin/io/waterkite94/stalk/usecase/port/SmtpPort.kt
@@ -1,0 +1,11 @@
+package io.waterkite94.stalk.usecase.port
+
+interface SmtpPort {
+    fun createVerificationCode(codeLength: Int): String
+
+    fun sendMailByJavaMailSender(
+        toEmail: String,
+        title: String,
+        content: String
+    )
+}
