+++ infrastructure/smtp/build.gradle.kts
@@ -1,3 +1,5 @@
 dependencies {
+    implementation(project(":application"))
+
     implementation("org.springframework.boot:spring-boot-starter-mail")
 }
