+++ infrastructure/smtp/src/main/resources/mail.yml
